% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convergence_path.R
\name{convergence_plot}
\alias{convergence_plot}
\title{show values by iteration}
\usage{
convergence_plot(df)
}
\arguments{
\item{df}{dataframe of convergence data from progress.txt}
}
\description{
show values by iteration
}

