% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cov_distributions.R
\name{cov_distributions}
\alias{cov_distributions}
\title{show covariate distributions}
\usage{
cov_distributions(df)
}
\arguments{
\item{df}{dataframe in format of StrCov pml output}
}
\description{
show covariate distributions
}
\details{
should have column names StrName (covariate name) and Covr (values)
}

