% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dv_vs_preds.R
\name{dv_vs_preds}
\alias{dv_vs_preds}
\title{create dv vs ipred or pred plots}
\usage{
dv_vs_preds(residual_df, pred_col = "IPRED")
}
\arguments{
\item{residual_df}{dataframe of residuals table}

\item{pred_col}{column name for pred column used to compare against DV}
}
\description{
create dv vs ipred or pred plots
}
\details{
residuals table found in out.txt of pml runs, or residual table in phoenix
 if multiple observations detected in ObsName column,
 will automatically facet by Obs
}

