% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eta_vs_cov.R
\name{eta_vs_cov}
\alias{eta_vs_cov}
\title{create Eta vs Covariate plots}
\usage{
eta_vs_cov(df)
}
\arguments{
\item{df}{dataframe of eta and covariate values from pml}
}
\description{
create Eta vs Covariate plots
}
\details{
expects 4 columns Covr (covariate values), Eta (eta values),
CovrName (covariate name), EtaName (name of eta)
can be accessed from pmldb from `.$EtaCov`
}

