% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/res_vs_ivar.R
\name{res_vs_ivar}
\alias{res_vs_ivar}
\title{create cwres vs ivar plots}
\usage{
res_vs_ivar(residual_df, res_col = "CWRES", ivar_col = "IVAR",
  central_tendency = list(method = "loess", group = NULL, size = 1.4, color =
  "red", se = FALSE), spread_tendency = list(method = "loess", group = NULL,
  size = 1, color = "blue", se = FALSE))
}
\arguments{
\item{residual_df}{dataframe of residuals table}

\item{res_col}{column name for residual column used}

\item{ivar_col}{column name for independent variable column used}

\item{central_tendency}{customization options for the central tendency}

\item{spread_tendency}{customization options for the absolute values of res data}
}
\description{
create cwres vs ivar plots
}
\details{
will automatically add a central and spread fits for the absolute value
 of the residual values to help indicate whether bias present. Default to
 loess curve, however can also use "method" = "lm" for a linear fit.
 To turn off either central or spread lines, set the argument to NULL
}

